Moved CommonTextField to core module
