+++ android/core/src/main/kotlin/com/hulkdx/findprofessional/core/commonui/CommonTextField.kt
@@ -1,4 +1,4 @@
-package com.hulkdx.findprofessional.feature.authentication.ui
+package com.hulkdx.findprofessional.core.commonui
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
@@ -12,11 +12,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.hulkdx.findprofessional.core.theme.AppTheme
 import com.hulkdx.findprofessional.core.theme.body2
-import com.hulkdx.findprofessional.feature.authentication.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -52,19 +49,4 @@ fun CommonTextField(
         ),
         visualTransformation = visualTransformation,
     )
-}
-
-@Preview
-@Composable
-private fun CommonTextFieldPreview() {
-    AppTheme {
-        CommonTextField(
-            modifier = Modifier,
-            hint = "hint",
-            leadingIconDrawable = R.drawable.ic_email,
-            leadingIconContentDescription = "",
-            value = "value",
-            onValueChanged = {},
-        )
-    }
 }
